certmanager: solve CannotRegenerateKey
`CertManager` might throw a warning with the current Helm chart because the `.spec.privateKey.rotationPolicy` is unset.

This will set the field to `Always` which allows rotation in case there is a need for, which can resolve potential issues of an invalid cert.

For context, this is what certmanager also recommends: https://cert-manager.io/docs/usage/certificate/

Signed-off-by: Samuel Lang <gh@lang-sam.de>